+++ src/Graphics/Vty/Output/Interface.hs
@@ -73,6 +73,8 @@ data Output = Output
       -- | Return the display to the state before `reserveDisplay` If no
       -- previous state then set the display state to the initial state.
     , releaseDisplay :: IO ()
+      -- | Sets the current display bounds (width, height).
+    , setDisplayBounds :: (Int, Int) -> IO ()
       -- | Returns the current display bounds.
     , displayBounds :: IO DisplayRegion
       -- | Output the bytestring to the terminal device.
@@ -99,6 +101,22 @@ data Output = Output
     , ringTerminalBell :: IO ()
       -- | Returns whether the terminal has an audio bell feature.
     , supportsBell :: IO Bool
+      -- | Returns whether the terminal supports italicized text.
+      --
+      -- This is terminal-dependent and should make a best effort to
+      -- determine whether this feature is supported, but even if the
+      -- terminal advertises support (e.g. via terminfo) that might not
+      -- be a reliable indicator of whether the feature will work as
+      -- desired.
+    , supportsItalics :: IO Bool
+      -- | Returns whether the terminal supports strikethrough text.
+      --
+      -- This is terminal-dependent and should make a best effort to
+      -- determine whether this feature is supported, but even if the
+      -- terminal advertises support (e.g. via terminfo) that might not
+      -- be a reliable indicator of whether the feature will work as
+      -- desired.
+    , supportsStrikethrough :: IO Bool
     }
 
 displayContext :: Output -> DisplayRegion -> IO DisplayContext